Module basilisk::service

Service and ServiceFactory implementation. Specialized wrapper over substrate service.

Structs

BasiliskExecutorDispatch

NewFull

TestingBasiliskExecutorDispatch

Traits

IdentifyVariant

Can be called for a Configuration to check what node it belongs to.

Functions

new_partial

new_partial_impl

parachain_build_import_queue

Build the import queue for the parachain runtime.

start_node

Start a normal parachain node.

start_node_impl

Start a node with the given parachain Configuration and relay chain Configuration.

Type Definitions

FullBackend

FullClient

Hash