Function basilisk::service::start_node_impl

async fn start_node_impl<RuntimeApi, Executor, BIC>(
    parachain_config: Configuration, 
    polkadot_config: Configuration, 
    para_id: ParaId, 
    build_consensus: BIC
) -> Result<NewFull<Arc<FullClient<RuntimeApi, Executor>>>> where
    RuntimeApi: ConstructRuntimeApi<Block, FullClient<RuntimeApi, Executor>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: RuntimeApiCollection<StateBackend = StateBackendFor<FullBackend, Block>>,
    RuntimeApi::RuntimeApi: CollectCollationInfo<Block>,
    RuntimeApi::RuntimeApi: AuraApi<Block, AuthorityId>,
    Executor: NativeExecutionDispatch + 'static,
    BIC: FnOnce(Arc<FullClient<RuntimeApi, Executor>>, Option<&Registry>, Option<TelemetryHandle>, &TaskManager, Arc<dyn RelayChainInterface>, Arc<FullPool<Block, FullClient<RuntimeApi, Executor>>>, Arc<NetworkService<Block, H256>>, SyncCryptoStorePtr, bool) -> Result<Box<dyn ParachainConsensus<Block>>, Error>, 

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.